refactor: unify pack and clean logic
